refactor: move filters to one package to meet template method design pattern
