+++ Design Document.md
@@ -13,9 +13,9 @@ users’ acceptable study hours.
 ##SOLID
 - Single responsibility principle:
   - SimpleScheduler: only responsible for scheduling the timetable via recursion. 
-  - Filter and its subclasses: For our three non-abstract filter classes, each of them is responsible for dealing with 
+  - Filter and its subclasses: For our three non-abstract ApplicationBusinessRule.filter classes, each of them is responsible for dealing with 
   one kind of user preference. For instance, Instructor Filter only works if users want to rule out specific unwanted 
-  instructors. Therefore, one non-abstract filter class has its own responsibility, and will change if and only its own filter method (named “sort” in the algorithms) changes.
+  instructors. Therefore, one non-abstract ApplicationBusinessRule.filter class has its own responsibility, and will change if and only its own ApplicationBusinessRule.filter method (named “sort” in the algorithms) changes.
   
 - Open/closed principle:
   - An implementation of the Open/closed principle is the Filter class. The instance variables in Filter are made 
@@ -25,7 +25,7 @@ users’ acceptable study hours.
 
 - Liskov Substitution principle:
   - For our Filter Class and its subclasses, all three subclasses of Filter only has an overwritten version the parent 
-  class's abstract sort() method, which is used to filter out the unwanted timetables based on users' preference. We did 
+  class's abstract sort() method, which is used to ApplicationBusinessRule.filter out the unwanted timetables based on users' preference. We did 
   not replace the parent class.
 
 ## Clean Architecture
@@ -41,7 +41,7 @@ users’ acceptable study hours.
   is the other member of Frameworks & Drivers. They cooperate to get information from both the user and the web 
   (In this case, it would be the UofT Coursefinder website). Then the controller calls the simple scheduler, giving it 
   the information from the user and the web. This way, it is clear who is doing what task. If the user has any 
-  preference, the controller learns this fact and hands it to the Filter. The filter determines what preference should 
+  preference, the controller learns this fact and hands it to the Filter. The ApplicationBusinessRule.filter determines what preference should 
   be applied. The finished product is in the form of a Timetable instance. Of course, the Timetable class also helps 
   this process by providing the state of itself. We achieved this by implementing helper methods in the Timetable class. 
   The data is handed over in the form of a Timetable to Javafx application for initial display to the user. 
@@ -52,10 +52,10 @@ users’ acceptable study hours.
 - In our project, we have implemented Template Method Design Pattern in the Filter Class and its subclasses. The Filter 
 is the base case that has constructor, call(), abstract sort(), getInput(), getOutput(), getUnwanted(), and 
 isFiltered() methods. The Filter class has three subclasses including Instructor Filter, Timeslot Filter and 
-Maximum Hour Filter. By running Filter class, the class will constructor a filter instance and call the call() method, 
+Maximum Hour Filter. By running Filter class, the class will constructor a ApplicationBusinessRule.filter instance and call the call() method, 
 which is a template method. If the user chooses Instructor Filter, the call() method will call InstructorFilter. 
 If the user chooses Timeslot Filter, the call() will call TimeslotFilter. If the user chooses Maximum Hour Filter, 
-the call() will call Maximum Hour Filter. The three child filter will rule out all the unwanted timetables based on the 
+the call() will call Maximum Hour Filter. The three child ApplicationBusinessRule.filter will rule out all the unwanted timetables based on the 
 users’ preference. The isFilter() method will check if the programs rule out some unwanted timetables by checking 
 the whether the output is empty. If the output is empty, the method will print out “Timetable are not filtered”. 
 Else, it will print out “Timetable are successfully filtered.” For #33 pull request, we started to implement the 
@@ -76,7 +76,7 @@ posted, we have added detail description to prevent confusion.
 course session instances. We also test that Timetable class and SimpleScheduler can successfully generate appropriate 
 timetables. Besides, we also tested the Instructor Filter, Timeslot Filter, and Maximum Hour Filter. Instructor Filter 
 can successfully rule out the sessions delivered by unwanted professors. Timeslot Filter can successfully rule out the 
-sessions with unwanted time slots. Maximum Hour Filter can successfully filter out the timetables that have lecture 
+sessions with unwanted time slots. Maximum Hour Filter can successfully ApplicationBusinessRule.filter out the timetables that have lecture 
 hours larger than the users’ preferential lecture hours.
 - The user interface functions appropriately on our computers.
 
@@ -127,13 +127,13 @@ application for initial display to the user.
   - Session is essentially the old Course class. It contains less information when compared to the old Course class for 
   the sake of efficiency. Session instances are stored in NewCourse.
   - Filter is an abstract class that will call different child classes based on the users' preference.
-  - Instructor filter: if you have any preference on instructors, for instance, you prefer not to enroll in specific
-      instructors’ sessions, please choose the Instructor Filter. With this filter, our program will automatically rule out
+  - Instructor ApplicationBusinessRule.filter: if you have any preference on instructors, for instance, you prefer not to enroll in specific
+      instructors’ sessions, please choose the Instructor Filter. With this ApplicationBusinessRule.filter, our program will automatically rule out
       the lecture sessions delivered by your unwanted instructors in your schedule. 
   - Maximum Hour Filter: if you have any preference on study hours per day, for instance, you prefer not to take more
   than 5 hours of lecture each day, please choose the Maximum Hour Filter. Our program will organize a timetable with an
   everyday lecture hour less or equal to your provided number. 
-  - Timeslot filter: if you have any preference on certain time slots, for instance, you prefer not to take a morning
+  - Timeslot ApplicationBusinessRule.filter: if you have any preference on certain time slots, for instance, you prefer not to take a morning
   course (e.g. 9 am – 10 am) or a night course (e.g. 8 pm – 10 pm), please choose the Timeslot Filter. We will rule out
   the sessions with your unwanted timeslots in your timetable.
 
@@ -144,7 +144,7 @@ every course will be stored in the same directory where our program is stored on
 the method. Everytime a user starts the program, our algorithms will automatically detect if the user has used the 
 program before by calling inputExist(). If our program detects that there is an existing version stored on the computer, 
 we would ask if you want to load your last version. If the user answers “Yes”, our program will load his or her stored 
-last version on the computer, and continue the rest process -- schedule timetables and filter unwanted timetables based 
+last version on the computer, and continue the rest process -- schedule timetables and ApplicationBusinessRule.filter unwanted timetables based 
 on the users’ preferences. The state persists across runs of our program.
